split fuelup check into command and ops
